requirements.txt
django
numpy
pandas
spylls
django-cors-headers
djangorestframework